Process output realtime streaming (#339)
* WIP: Prepare message for streaming

* Add stream flag in command

* Implement message forwarding server side

* Implement client side stream method

* Update vendor

* improvement in stream method

Avoid blocking by checking that the process is running or it
has some buffered output

* Update docs

* Allow None streams

* Make sure to decode before deserialize

* Fix typos

+++ base/pm/core/result.go
@@ -38,7 +38,7 @@ type JobResult struct {
 	Data      string  `json:"data"`
 	Streams   Streams `json:"streams,omitempty"`
 	Critical  string  `json:"critical,omitempty"`
-	Level     int     `json:"level"`
+	Level     uint16  `json:"level"`
 	State     string  `json:"state"`
 	StartTime int64   `json:"starttime"`
 	Time      int64   `json:"time"`
